requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
filelock==3.18.0
fsspec==2025.3.2
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
openai-whisper==20240930
portalocker==2.10.1
protobuf==6.31.0
pydantic==2.11.4
pydantic_core==2.33.2
qdrant-client==1.14.2
regex==2024.11.6
requests==2.32.3
setuptools==80.7.1
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
yt-dlp==2025.4.30
